Initial SplitView refactor
