+++ ReaderIOS/ReaderIOS/Views/PDFView.swift
@@ -455,5 +455,5 @@ struct MarkupMenu: View {
 }
 
 #Preview {
-    NavigationPDFSplitView()
+    SplitView()
 }
