+++ ReaderIOS/ReaderIOS/Views/SplitViewComponents/SplitView.swift
@@ -1,0 +1,191 @@
+//
+//  TableOfContentsSplitView.swift
+//  ReaderIOS
+//
+//  Created by Devin Hadley on 11/10/24.
+//
+import PDFKit
+import SwiftUI
+
+struct SplitView: View {
+    // Loaded workbooks information state vars
+    @State private var workbooks: [Workbook]?
+    @State private var chapters: [Chapter]?
+    @State private var covers: [Cover]?
+
+    // User selection (what they are viewing) state vars
+    @State private var selectedWorkbookID: String?
+    @State private var selectedChapterID: String?
+    @State private var currentPage: Int = 0
+    @State private var currentPdfFileName: String?
+
+    @State private var columnVisibility = NavigationSplitViewVisibility.automatic
+
+    // Bookmark state vars
+    @State private var isShowingBookmarks: Bool = false
+    @State private var bookmarkLookup = [String: Set<Int>]()
+
+    // PDFDocument loaded by PDFView
+    @State private var pdfDocument: PDFDocument?
+
+    // Chapter manager, detemines chapter from current page
+    @State private var chapterManager: ChapterManager?
+
+    var body: some View {
+        NavigationSplitView(columnVisibility: $columnVisibility) {
+            if let workbooks = workbooks {
+                List(workbooks, selection: $selectedWorkbookID) { workbook in
+                    HStack {
+                        Image(systemName: "icloud.and.arrow.down") // Download icon
+                            .font(.caption)
+                            .foregroundColor(.blue)
+                        Text(workbook.id)
+                            .tag(workbook.id)
+                    }
+                }
+            } else {
+                ProgressView("Fetching Workbooks")
+                    .onAppear {
+                        fetchWorkbooks()
+                    }
+            }
+        }
+        content: {
+            Group {
+                if !isShowingBookmarks {
+                    // Search view
+                    SearchView(
+                        currentPage: $currentPage,
+                        columnVisibility: $columnVisibility,
+                        chapters: chapters,
+                        fetchChapters: fetchChapters,
+                        pdfDocument: pdfDocument
+                    )
+                } else {
+                    // Bookmarks view
+                    BookmarkView(
+                        currentPage: $currentPage,
+                        currentPdfFileName: currentPdfFileName,
+                        bookmarkLookup: bookmarkLookup
+                    )
+                }
+            }
+            .toolbar {
+                ToolbarItem(placement: .automatic) {
+                    Toggle(isOn: $isShowingBookmarks) {
+                        Image(systemName: isShowingBookmarks ? "bookmark.fill" : "bookmark")
+                            .foregroundColor(.accentColor)
+                    }
+                    .toggleStyle(.button)
+                    .buttonStyle(.plain)
+                    .accessibilityLabel(isShowingBookmarks ? "Show All Chapters" : "Show Bookmarked Chapters")
+                }
+            }
+        } detail: {
+            if currentPdfFileName != nil {
+                // TODO: Only give access to bookmarks for current file.
+                PDFView(
+                    fileName: $currentPdfFileName,
+                    currentPage: $currentPage,
+                    bookmarkLookup: $bookmarkLookup,
+                    covers: $covers,
+                    pdfDocument: $pdfDocument
+                )
+            } else {
+                ProgressView("Getting the latest workbook.")
+            }
+        }
+        .onChange(of: selectedWorkbookID) {
+            guard let selectedWorkbook = selectedWorkbook else { return }
+
+            if currentPdfFileName != selectedWorkbook.pdfName {
+                currentPdfFileName = selectedWorkbook.pdfName
+            }
+
+            fetchChapters()
+
+            if let selectedWorkbookID {
+                currentPage = StateRestoreManager.shared.loadPageNumber(for: selectedWorkbookID)
+            }
+
+            persistState()
+        }
+        .onChange(of: currentPage) { _, _ in
+            updateSelectedChapter()
+
+            persistState()
+        }
+    }
+
+    var selectedWorkbook: Workbook? {
+        workbooks?.first(where: { $0.id == selectedWorkbookID })
+    }
+
+    func persistState() {
+        if let selectedWorkbookID {
+            StateRestoreManager.shared.saveState(workbookID: selectedWorkbookID, pageNumber: currentPage)
+        }
+    }
+
+    func fetchWorkbooks() {
+        NetworkingService.shared.fetchWorkbooks { result in
+            switch result {
+            case let .success(workbookResponse):
+                workbooks = workbookResponse
+
+                // Try to load saved state now that workbooks are available.
+                if let savedState = StateRestoreManager.shared.loadState() {
+                    if workbookResponse.first(where: { $0.id == savedState.workbookID }) != nil {
+                        selectedWorkbookID = savedState.workbookID
+                        currentPage = savedState.pageNumber
+                    } else {
+                        // Fallback: default to the first workbook if the saved one isn't found.
+                        selectedWorkbookID = workbookResponse.first?.id
+                    }
+                } else {
+                    // No saved state, so select the first workbook by default.
+                    selectedWorkbookID = workbookResponse.first?.id
+                }
+            case let .failure(error):
+                print("Error fetching workbooks: \(error)")
+            }
+        }
+    }
+
+    func fetchChapters() {
+        guard let fileName = selectedWorkbook?.metaName else { return }
+
+        NetworkingService.shared.fetchChapters(metaName: fileName) { result in
+            switch result {
+            case let .success(chapterResponse):
+                chapters = chapterResponse
+                selectedChapterID = chapters?.first?.id
+                setupChapterManager()
+            case let .failure(error):
+                print("Error fetching chapters: \(error)")
+            }
+        }
+    }
+
+    func setupChapterManager() {
+        if let chapters = chapters {
+            let chapterManager = ChapterManager(chapters: chapters)
+            // Update selectedChapterID whenever currentPage changes
+            self.chapterManager = chapterManager
+        }
+    }
+
+    func updateSelectedChapter() {
+        if let chapterManager = chapterManager {
+            if let newChapter = chapterManager.getChapter(forPage: currentPage) {
+                selectedChapterID = newChapter.id
+                covers = newChapter.covers
+                print("Updated covers: \(covers?.map(\.desc) ?? [])")
+            }
+        }
+    }
+}
+
+#Preview {
+    SplitView()
+}
